fix(link-item): move LinkItem to /components folder
fix(link-item): move LinkItem to /components folder
